Add custom dropdown directive to course project
This custom dropdown directive replaces functionality provided
out-of-the-box by bootstrap4, allowing a dropdown button to be toggled
by a click. In a real-world scenario, I'd use the functionality provided
by bootstrap, but since this course is on angular, I guess it makes
sense to learn how to implement all dynamic actions with angular.